projectfinal/requirements.txt
langchain_google_genai
streamlit
chromadb
psycopg2

speech_recognition